refactor(llm): add shared LLM utility and module-level analyst prompt; refactor analyst to use it
